Separating components. Commenting
