+++ react-app/src/pages/UsersTable.tsx
@@ -3,171 +3,22 @@ import { CustomDataGrid } from '@/components/table/DataTable';
 import {
   Box,
   Chip,
-  InputAdornment,
   Paper,
   SxProps,
-  TextField,
   Typography,
   debounce,
   useTheme,
   IconButton,
 } from '@mui/material';
 import { GridColDef, useGridApiRef } from '@mui/x-data-grid';
-import React, { Dispatch, SetStateAction, useEffect, useState } from 'react';
-import { UUID } from 'crypto';
+import React, { useEffect, useState } from 'react';
 import { useKeycloak } from '@bcgov/citz-imb-kc-react';
-import SearchIcon from '@mui/icons-material/Search';
-import CloseIcon from '@mui/icons-material/Close';
 import FilterAltOffIcon from '@mui/icons-material/FilterAltOff';
-
-// interface IAgency {
-//   createdOn: string;
-//   updatedOn: string;
-//   updatedByName: string;
-//   updatedByEmail: string;
-//   id: number;
-//   name: string;
-//   isDisabled: true;
-//   isVisible: true;
-//   sortOrder: number;
-//   type: string;
-//   code: string;
-//   parentId: number;
-//   description: string;
-// }
-
-// interface IRole {
-//   createdOn: string;
-//   updatedOn: string;
-//   updatedByName: string;
-//   updatedByEmail: string;
-//   id: UUID;
-//   name: string;
-//   isDisabled: true;
-//   isVisible: true;
-//   sortOrder: 0;
-//   type: string;
-//   description: string;
-//   keycloakGroupId: UUID;
-//   isPublic: true;
-// }
-
-interface IUser {
-  createdOn: string;
-  updatedOn: string;
-  updatedByName: string;
-  updatedByEmail: string;
-  id: UUID;
-  keycloakid: UUID;
-  username: string;
-  position: string;
-  displayName: string;
-  firstName: string;
-  middleName: string;
-  lastName: string;
-  email: string;
-  isDisabled: true;
-  emailVerified: true;
-  note: string;
-  lastLogin: string;
-  agency: string;
-  roles: string;
-}
-
-interface IKeywordSearchProps {
-  onChange?: Function;
-  optionalExternalState?: [string, Dispatch<SetStateAction<string>>];
-}
-
-const KeywordSearch = (props: IKeywordSearchProps) => {
-  const { onChange, optionalExternalState } = props;
-  const [isOpen, setIsOpen] = useState<boolean>(false);
-  const [fieldContents, setFieldContents] = optionalExternalState
-    ? optionalExternalState
-    : useState<string>('');
-  const theme = useTheme();
-
-  const commonStyle: SxProps = {
-    fontSize: theme.typography.fontWeightBold,
-    fontFamily: theme.typography.fontFamily,
-    padding: '5px',
-    marginBottom: '1px',
-    boxSizing: 'content-box',
-    borderRadius: '5px',
-  };
-
-  const openStyle: SxProps = {
-    ...commonStyle,
-    width: '240px',
-    transition: 'width 0.3s ease-in, border 1s',
-    border: `1.5px solid ${theme.palette.grey[400]}`,
-    '&:focus-within': {
-      border: '1.5px solid black',
-    },
-  };
-
-  const closedStyle: SxProps = {
-    ...commonStyle,
-    width: '32px',
-    transition: 'width 0.3s ease-in, border 1s',
-    border: '1.5px solid transparent',
-    '&:hover': {
-      cursor: 'default',
-    },
-  };
-  return (
-    <TextField
-      id="keyword-search"
-      variant="standard"
-      sx={isOpen ? openStyle : closedStyle}
-      size="small"
-      style={{ fontSize: '5em' }}
-      placeholder="Search..."
-      value={fieldContents}
-      onChange={(e) => {
-        setFieldContents(e.target.value);
-        onChange(e.target.value);
-      }}
-      InputProps={{
-        disableUnderline: true,
-        sx: { cursor: 'default' },
-        endAdornment: (
-          <InputAdornment position={'end'}>
-            {fieldContents ? (
-              <CloseIcon
-                onClick={() => {
-                  // Clear text and filter
-                  setFieldContents('');
-                  onChange('');
-                  document.getElementById('keyword-search').focus();
-                }}
-                sx={{
-                  '&:hover': {
-                    cursor: 'pointer',
-                  },
-                }}
-              />
-            ) : (
-              <SearchIcon
-                onClick={() => {
-                  setIsOpen(!isOpen);
-                  document.getElementById('keyword-search').focus();
-                }}
-                sx={{
-                  '&:hover': {
-                    cursor: 'pointer',
-                  },
-                }}
-              />
-            )}
-          </InputAdornment>
-        ),
-      }}
-    />
-  );
-};
+import KeywordSearch from '@/components/table/KeywordSearch';
+import { IUser } from '@/interfaces/IUser';
 
 const UsersTable = () => {
+  // States and contexts
   const [users, setUsers] = useState<IUser[]>([]);
   const [rowCount, setRowCount] = useState<number>(0);
   const [keywordSearchContents, setKeywordSearchContents] = useState<string>('');
@@ -178,6 +29,7 @@ const UsersTable = () => {
 
   useEffect(() => {
     if (state.accessToken) {
+      // FIXME: This should be using the proxy...
       fetch('http://localhost:5000/api/v2/admin/users', {
         headers: {
           Authorization: getAuthorizationHeaderValue(),
@@ -190,23 +42,26 @@ const UsersTable = () => {
           }
         })
         .catch((e: unknown) => {
-          console.log(e);
+          console.warn(e);
         });
     }
   }, [state]);
 
+  // Determines colours of chips
   const colorMap = {
     Pending: 'info',
     Active: 'success',
     Hold: 'warning',
   };
 
+  // Sets quickfilter value of DataGrid. newValue is a string input.
   const updateSearchValue = React.useMemo(() => {
     return debounce((newValue) => {
       apiRef.current.setQuickFilterValues(newValue.split(' ').filter((word) => word !== ''));
     }, 100);
   }, [apiRef]);
 
+  // Converts dates to the MMM DD, YYYY locale
   const dateFormatter = (params) =>
     new Intl.DateTimeFormat('en-US', {
       month: 'short',
@@ -214,6 +69,7 @@ const UsersTable = () => {
       year: 'numeric',
     }).format(new Date(params.value));
 
+  // Defines the columns used in the table.
   const columns: GridColDef[] = [
     {
       field: 'firstName',
@@ -315,6 +171,7 @@ const UsersTable = () => {
               {keywordSearchContents || gridFilterItems.length > 0 ? (
                 <IconButton
                   onClick={() => {
+                    // Set both DataGrid and Keyword search back to blanks
                     apiRef.current.setFilterModel({ items: [] });
                     setKeywordSearchContents('');
                   }}
@@ -325,7 +182,6 @@ const UsersTable = () => {
                 <></>
               )}
             </Box>
-
             <KeywordSearch
               onChange={updateSearchValue}
               optionalExternalState={[keywordSearchContents, setKeywordSearchContents]}
@@ -336,11 +192,13 @@ const UsersTable = () => {
             columns={columns}
             rows={users}
             onStateChange={(e) => {
+              // Keep track of row count separately
               setRowCount(
                 Object.values(e.filter.filteredRowsLookup).filter((value) => value).length,
               );
             }}
             onFilterModelChange={(e) => {
+              // Get the filter items from MUI, filter out blanks, set state
               setGridFilterItems(e.items.filter((item) => item.value));
             }}
             apiRef={apiRef}
